fix(swig): spell H2MM/BVA burst inputs as vector<long long> for Linux
On LP64 Linux the Python wrapper is generated with -DSWIGWORDSIZE64 and
without stdint.i (deliberate, see misc_types.i), so SWIG keeps int64_t
literal and gives std::vector<int64_t> elements only opaque
pointer-category traits: a Python list of ints cannot convert, and
BVA.compute / H2MM.set_bursts_from_tttr rejected the documented
'[s0, e0, s1, e1, ...]' list input (TypeError on every Linux CI test
job). macOS/Windows resolve int64_t to long long and were unaffected.

Spell the public burst-list parameters std::vector<long long> - the
same convention TTTR::burst_search already uses; vector<long long> has
real value traits in every wrapper. Internal storage stays int64_t.

+++ include/BVA.h
@@ -73,9 +73,13 @@ public:
      *        consecutive chunks of this many photons.  If <= 0, slice into
      *        fixed-duration time windows of ``minimum_window_length`` seconds.
      * @param minimum_window_length Slice duration in seconds (time-window mode).
+     *
+     * The burst vector is spelled std::vector<long long> (not int64_t): SWIG
+     * treats int64_t elements as opaque in std_vector on LP64 Linux, so a
+     * Python list would not convert. Same convention as TTTR::burst_search.
      */
     void compute(
-        const std::vector<int64_t>& bursts,
+        const std::vector<long long>& bursts,
         int number_of_photons_per_slice = -1,
         double minimum_window_length = 0.01
     );
